expire mlinz snapshots

+++ expire-zfs-snapshots.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -xe
+#!/bin/sh -e
 
 keep=14
 
@@ -10,10 +10,23 @@ zfs_destroy() {
 	echo -n "$name snapshots: "
 	wc -l /tmp/zfs.$name
 	tail -$keep /tmp/zfs.$name > /tmp/zfs.$name.keep
-	grep -v -f /tmp/zfs.$name.keep /tmp/zfs.$name > /tmp/zfs.$name.destroy
-	echo -n "$name destroy:   "
-	wc -l /tmp/zfs.$name.destroy
-	cat /tmp/zfs.$name.destroy | xargs -i zfs destroy -v {}
+	if grep -v -f /tmp/zfs.$name.keep /tmp/zfs.$name > /tmp/zfs.$name.destroy ; then
+		echo -n "$name destroy:   "
+		wc -l /tmp/zfs.$name.destroy
+		cat /tmp/zfs.$name.destroy | xargs -i zfs destroy -v {}
+	else
+		echo "$name no snapshots older than $keep days"
+	fi
 }
 
 zfs_destroy zamd/dpavlin/nuc
+
+zfs_destroy zamd/mglavica/mlinz/data
+zfs_destroy zamd/mglavica/mlinz/data/rest
+zfs_destroy zamd/mglavica/mlinz/docker
+zfs_destroy zamd/mglavica/mlinz/home
+zfs_destroy zamd/mglavica/mlinz/home/mglavica
+zfs_destroy zamd/mglavica/mlinz/klin
+zfs_destroy zamd/mglavica/mlinz/postgresql
+zfs_destroy zamd/mglavica/mlinz/root
+zfs_destroy zamd/mglavica/mlinz/srv
